@@ -1,4 +1,4 @@
-package com.example.myapplication;
+package com.aayush.signinapp;
 
 import org.junit.Test;
 
@@ -9,7 +9,7 @@ import static org.junit.Assert.*;
  *
  * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
  */
-public class ExampleUnitTest {
+public class UnitTest {
     @Test
     public void addition_isCorrect() {
         assertEquals(4, 2 + 2);
